[lp-2189] : create generalpartner package and move classes into it
