+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/generalpartner/dao/GeneralPartnerDataDao.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.limitedpartnershipsapi.model.generalpartner.dao;
+package uk.gov.companieshouse.limitedpartnershipsapi.generalpartner.dao;
 
 import org.springframework.data.mongodb.core.mapping.Field;
 import uk.gov.companieshouse.limitedpartnershipsapi.model.common.dao.PartnerDataDao;
