+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/generalpartner/dto/GeneralPartnerDataDto.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.limitedpartnershipsapi.model.generalpartner.dto;
+package uk.gov.companieshouse.limitedpartnershipsapi.generalpartner.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
