+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/generalpartner/dto/GeneralPartnerSubmissionCreatedResponseDto.java
@@ -1,0 +1,4 @@
+package uk.gov.companieshouse.limitedpartnershipsapi.generalpartner.dto;
+
+public record GeneralPartnerSubmissionCreatedResponseDto(String id) {
+}
