--- src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/model/generalpartner/dto/GeneralPartnerSubmissionCreatedResponseDto.java
@@ -1,4 +1,0 @@
-package uk.gov.companieshouse.limitedpartnershipsapi.model.generalpartner.dto;
-
-public record GeneralPartnerSubmissionCreatedResponseDto(String id) {
-}
